stats & memory leak fixed?

+++ src/com/Nips/TnTSlap/Stats/ValueComparator.java
@@ -5,9 +5,9 @@ import java.util.Map;
 
 public class ValueComparator implements Comparator<String> {
 
-	Map<String, Double> base;
+	Map<String, Integer> base;
 
-	public ValueComparator(Map<String, Double> base) {
+	public ValueComparator(Map<String, Integer> base) {
 		this.base = base;
 	}
